+++ src/test/java/io/github/astrapi69/xml/to/xsd/XmlToXsdExtensionsTest.java
@@ -42,6 +42,10 @@ import io.github.astrapi69.checksum.FileChecksumExtensions;
 import io.github.astrapi69.collections.array.ArrayFactory;
 import io.github.astrapi69.file.delete.DeleteFileExtensions;
 import io.github.astrapi69.file.search.PathFinder;
+import io.github.astrapi69.test.object.Employee;
+import io.github.astrapi69.test.object.Person;
+import io.github.astrapi69.test.object.enumtype.Gender;
+import io.github.astrapi69.xml.ObjectToXmlExtensions;
 
 /**
  * The unit test class for the class {@link XmlToXsdExtensions}
@@ -49,14 +53,41 @@ import io.github.astrapi69.file.search.PathFinder;
 public class XmlToXsdExtensionsTest
 {
 
-	/**
-	 * Test method for {@link XmlToXsdExtensions}
-	 */
 	@Test
-	public void testWithBeanTester()
+	public void testObjectToXmlToXsd() throws IOException, XmlException
 	{
-		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(XmlToXsdExtensions.class);
+		String actual;
+		String expected;
+		Person person;
+		Employee employee;
+
+		person = Person.builder().gender(Gender.FEMALE).name("Anna").nickname(null).married(null)
+			.about(null).build();
+
+		employee = Employee.builder().id("23").person(person).build();
+
+		String xml = ObjectToXmlExtensions.toXmlWithJackson(employee);
+
+		actual = XmlToXsdExtensions.xmlToXsd(xml);
+		expected = "<xs:schema attributeFormDefault=\"unqualified\" elementFormDefault=\"qualified\" xmlns:xs=\"http://www.w3.org/2001/XMLSchema\">\n"
+			+ "  <xs:element name=\"Employee\" type=\"EmployeeType\"/>\n"
+			+ "  <xs:complexType name=\"personType\">\n" + "    <xs:sequence>\n"
+			+ "      <xs:element type=\"xs:string\" name=\"about\"/>\n"
+			+ "      <xs:element type=\"xs:string\" name=\"gender\"/>\n"
+			+ "      <xs:element type=\"xs:string\" name=\"married\"/>\n"
+			+ "      <xs:element type=\"xs:string\" name=\"name\"/>\n"
+			+ "      <xs:element type=\"xs:string\" name=\"nickname\"/>\n" + "    </xs:sequence>\n"
+			+ "  </xs:complexType>\n" + "  <xs:complexType name=\"EmployeeType\">\n"
+			+ "    <xs:sequence>\n" + "      <xs:element type=\"xs:byte\" name=\"id\"/>\n"
+			+ "      <xs:element type=\"personType\" name=\"person\"/>\n" + "    </xs:sequence>\n"
+			+ "  </xs:complexType>\n" + "</xs:schema>";
+		assertEquals(actual, expected);
+		employee = Employee.builder().id("1").person(Person.builder().build()).build();
+
+		xml = ObjectToXmlExtensions.toXmlWithJackson(employee);
+
+		actual = XmlToXsdExtensions.xmlToXsd(xml);
+		assertEquals(actual, expected);
 	}
 
 	/**
@@ -334,6 +365,17 @@ public class XmlToXsdExtensionsTest
 			+ "      <xs:element type=\"xs:string\" name=\"name\"/>\n" + "    </xs:sequence>\n"
 			+ "  </xs:complexType>\n" + "</xs:schema>";
 		assertEquals("Expected should be equal with the result.", expected, actual);
+	}
+
+
+	/**
+	 * Test method for {@link XmlToXsdExtensions}
+	 */
+	@Test
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(XmlToXsdExtensions.class);
 	}
 
 }